requirements.txt
absl-py==2.3.0
altair==6.0.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.2.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
curl_cffi==0.11.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.14
decorator==5.2.1
Django==5.2
executing==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
finnhub-python==2.4.25
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.1
FreeSimpleGUI==5.1.1
frozendict==2.4.6
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
immutabledict==4.2.1
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multitasking==0.0.11
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.1.0
ortools==9.12.4544
packaging==25.0
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.5.0
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
rpds-py==0.30.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tkcalendar==1.6.1
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchsummary==1.5.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.65
